+++ R/Filter.R
@@ -294,11 +294,8 @@ randomForestSRC.var.select = makeFilter( # nolint
     im = randomForestSRC::var.select(getTaskFormula(task), getTaskData(task),
       method = method, verbose = FALSE, always.use = getTaskFeatureNames(task),
       ...)
-    if (method == "md") {
-      setNames(im$varselect$depth, im$topvars)
-    } else if (method == "vh") {
-      im[["varselect"]][, 1L]
-    }
+    im$varselect[setdiff(rownames(im$varselect), im$topvars), "depth"] = NA
+    im$topvars
   })
 .FilterRegister[["randomForestSRC.var.select"]] = randomForestSRC.var.select
 .FilterRegister[["randomForestSRC.var.select"]]$desc = "(DEPRECATED)"